- catalog re-design - hide keyboard review fix
