--- iOS-TrackerApp/iOS-TrackerApp/ViewControllers/TrackerFormViewController.swift
@@ -1,317 +1,0 @@
-//
-//  TrackerFormViewController.swift
-//  iOS-TrackerApp
-//
-//  Created by Andy Kruch on 31.07.23.
-//
-
-import UIKit
-
- protocol TrackerFormViewControllerDelegate: AnyObject {
-     func didTapCancelButton()
-     func didTapConfirmButton(categoryLabel: String, trackerToAdd: Tracker)
- }
-
- final class TrackerFormViewController: UIViewController {
-     // MARK: - Layout elements
-     
-     private lazy var textField: UITextField = {
-         let textField = TextField(placeholder: "Введите название трекера")
-         textField.addTarget(self, action: #selector(didChangedLabelTextField), for: .editingChanged)
-         return textField
-     }()
-     
-     private let validationMessage: UILabel = {
-         let label = UILabel()
-         label.font = UIFont.systemFont(ofSize: 17)
-         label.textColor = .Red
-         label.text = "Ограничение 38 символов"
-         return label
-     }()
-     
-     private let parametersTableView: UITableView = {
-         let table = UITableView()
-         table.separatorStyle = .none
-         table.isScrollEnabled = false
-         table.register(ListCell.self, forCellReuseIdentifier: ListCell.identifier)
-         return table
-     }()
-     
-     private lazy var cancelButton: UIButton = {
-         let button = makeButton()
-         button.setTitle("Отменить", for: .normal)
-         button.setTitleColor(.red, for: .normal)
-         button.backgroundColor = .white
-         button.layer.borderWidth = 1
-         button.layer.borderColor = UIColor.red.cgColor
-         button.addTarget(self, action: #selector(didTapCancelButton), for: .touchUpInside)
-         return button
-     }()
-     
-     private lazy var confirmButton: UIButton = {
-         let button = makeButton()
-         button.setTitle("Создать", for: .normal)
-         button.setTitleColor(.WhiteDay, for: .normal)
-         button.backgroundColor = .Gray
-         button.addTarget(self, action: #selector(didTapConfirmButton), for: .touchUpInside)
-         button.isEnabled = false
-         return button
-     }()
-     
-     private let buttonsStack: UIStackView = {
-         let stack = UIStackView()
-         stack.spacing = 8
-         stack.distribution = .fillEqually
-         return stack
-     }()
-
-     // MARK: - Properties
-     
-     weak var delegate: TrackerFormViewControllerDelegate?
-     private let type: SetTrackersViewController.TrackerType
-     private var data: Tracker.Data {
-         didSet {
-             checkFromValidation()
-         }
-     }
-     
-     private var category: String? = TrackerCategory.mockData[0].label {
-         didSet {
-             checkFromValidation()
-         }
-     }
-     
-     private var scheduleString: String? {
-         guard let schedule = data.schedule else { return nil }
-         if schedule.count == WeekDay.allCases.count { return "Каждый день" }
-         let shortForms: [String] = schedule.map { $0.shortForm }
-         return shortForms.joined(separator: ", ")
-     }
-     
-     private var isConfirmButtonEnabled: Bool = false {
-         willSet {
-             if newValue {
-                 confirmButton.backgroundColor = .BlackDay
-                 confirmButton.isEnabled = true
-             } else {
-                 confirmButton.backgroundColor = .gray
-                 confirmButton.isEnabled = false
-             }
-         }
-     }
-     
-     private var isValidationMessageVisible = false {
-         didSet {
-             checkFromValidation()
-             if isValidationMessageVisible {
-                 validationMessageHeightConstraint?.constant = 22
-                 parametersTableViewTopConstraint?.constant = 32
-             } else {
-                 validationMessageHeightConstraint?.constant = 0
-                 parametersTableViewTopConstraint?.constant = 16
-             }
-         }
-     }
-     
-     private var validationMessageHeightConstraint: NSLayoutConstraint?
-     private var parametersTableViewTopConstraint: NSLayoutConstraint?
-     private let parameters = ["Категория", "Расписание"]
-     private let emojis = emojisArray
-     private let colors = UIColor.bunchOfSChoices
-
-     // MARK: - Lifecycle
-     init(type: SetTrackersViewController.TrackerType, data: Tracker.Data = Tracker.Data()) {
-         self.type = type
-         self.data = data
-         switch type {
-         case .habit:
-             self.data.schedule = []
-         case .irregularEvent:
-             self.data.schedule = nil
-         }
-         super.init(nibName: nil, bundle: nil)
-     }
-
-     required init?(coder: NSCoder) {
-         fatalError("init(coder:) has not been implemented")
-     }
-
-     private var collectionViewHeightConstraint: NSLayoutConstraint!
-
-     override func viewDidLoad() {
-         super.viewDidLoad()
-
-         configureViews()
-         configureConstraints()
-
-         data.emoji = emojis.randomElement()
-         data.color = colors.randomElement()
-
-         checkFromValidation()
-     }
-
-     // MARK: - Actions
-     
-     @objc
-     private func didChangedLabelTextField(_ sender: UITextField) {
-         guard let text = sender.text else { return }
-         data.label = text
-         if text.count > 38 {
-             isValidationMessageVisible = true
-         } else {
-             isValidationMessageVisible = false
-         }
-     }
-
-     @objc
-     private func didTapCancelButton() {
-         delegate?.didTapCancelButton()
-     }
-
-     @objc
-     private func didTapConfirmButton() {
-         guard let category, let emoji = data.emoji, let color = data.color else { return }
-
-         let newTracker = Tracker(
-             label: data.label,
-             emoji: emoji,
-             color: color,
-             schedule: data.schedule
-         )
-         delegate?.didTapConfirmButton(categoryLabel: category, trackerToAdd: newTracker)
-     }
-
-     // MARK: - Methods
-     private func checkFromValidation() {
-         if data.label.count == 0 {
-             isConfirmButtonEnabled = false
-             return
-         }
-         if isValidationMessageVisible {
-             isConfirmButtonEnabled = false
-             return
-         }
-         if category == nil || data.emoji == nil || data.color == nil {
-             isConfirmButtonEnabled = false
-             return
-         }
-         if let schedule = data.schedule, schedule.isEmpty {
-             isConfirmButtonEnabled = false
-             return
-         }
-         isConfirmButtonEnabled = true
-     }
-     
-     private func makeButton() -> UIButton {
-         let button = UIButton()
-         button.translatesAutoresizingMaskIntoConstraints = false
-         button.titleLabel?.font = UIFont.systemFont(ofSize: 16)
-         button.layer.cornerRadius = 16
-         button.layer.masksToBounds = true
-         return button
-     }
- }
-
- // MARK: - Layout methods
- private extension TrackerFormViewController {
-     func configureViews() {
-         switch type {
-         case .habit: title = "Новая привычка"
-         case .irregularEvent: title = "Новое нерегулярное событие"
-         }
-
-         parametersTableView.dataSource = self
-         parametersTableView.delegate = self
-
-         view.backgroundColor = .WhiteDay
-         
-         [textField, validationMessage, parametersTableView, buttonsStack].forEach { view.addSubview($0) }
-         
-         buttonsStack.addArrangedSubview(cancelButton)
-         buttonsStack.addArrangedSubview(confirmButton)
-         
-         textField.translatesAutoresizingMaskIntoConstraints = false
-         validationMessage.translatesAutoresizingMaskIntoConstraints = false
-         parametersTableView.translatesAutoresizingMaskIntoConstraints = false
-         buttonsStack.translatesAutoresizingMaskIntoConstraints = false
-     }
-
-     func configureConstraints() {
-         validationMessageHeightConstraint = validationMessage.heightAnchor.constraint(equalToConstant: 0)
-         parametersTableViewTopConstraint = parametersTableView.topAnchor.constraint(equalTo: validationMessage.bottomAnchor, constant: 16)
-         validationMessageHeightConstraint?.isActive = true
-         parametersTableViewTopConstraint?.isActive = true
-
-         NSLayoutConstraint.activate([
-             textField.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 16),
-             textField.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 24),
-             textField.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor, constant: -16),
-             textField.heightAnchor.constraint(equalToConstant: 75),
-             validationMessage.centerXAnchor.constraint(equalTo: textField.centerXAnchor),
-             validationMessage.topAnchor.constraint(equalTo: textField.bottomAnchor, constant: 8),
-             parametersTableView.leadingAnchor.constraint(equalTo: textField.leadingAnchor),
-             parametersTableView.trailingAnchor.constraint(equalTo: textField.trailingAnchor),
-             parametersTableView.heightAnchor.constraint(equalToConstant: data.schedule == nil ? ListOfItems.height : 2 *  ListOfItems.height),
-             buttonsStack.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 20),
-             buttonsStack.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor, constant: -20),
-             buttonsStack.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor),
-             buttonsStack.heightAnchor.constraint(equalToConstant: 60)
-         ])
-     }
- }
-
- // MARK: - UITableViewDataSource
- extension TrackerFormViewController: UITableViewDataSource {
-     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-         if data.schedule == nil {
-             return 1
-         }
-         return 2
-     }
-
-     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-         guard let listCell = tableView.dequeueReusableCell(withIdentifier: ListCell.identifier) as? ListCell
-         else { return UITableViewCell() }
-
-         var position: ListOfItems.Position
-         var value: String? = nil
-
-         if data.schedule == nil {
-             position = .alone
-             value = category
-         } else {
-             position = indexPath.row == 0 ? .first : .last
-             value = indexPath.row == 0 ? category : scheduleString
-         }
-
-         listCell.configure(label: parameters[indexPath.row], value: value, position: position)
-         return listCell
-     }
- }
-
- // MARK: - UITableViewDelegate
-extension TrackerFormViewController: UITableViewDelegate {
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        switch indexPath.row {
-        case 1:
-            guard let schedule = data.schedule else { return }
-            let scheduleViewController = ScheduleViewController(selectedWeekdays: schedule)
-            scheduleViewController.delegate = self
-            let navigationController = UINavigationController(rootViewController: scheduleViewController)
-            present(navigationController, animated: true)
-        default:
-            return
-        }
-    }
-        func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-            ListOfItems.height
-        }
-    }
-    
-    extension TrackerFormViewController: ScheduleViewControllerDelegate {
-        func didConfirm(_ schedule: [WeekDay]) {
-            data.schedule = schedule
-            parametersTableView.reloadData()
-            dismiss(animated: true)
-        }
-    }
